pyproject.toml
[tool.poetry]
name = "schedule"
version = "0.1.0"
description = ""
authors = ["steven"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
